requirements.txt
discord.py
flask
yt_dlp
